#23 refactor: 피드 목록 뷰, 데이터 로직 분리
